AzureSignTool emits no cert expiration date. Don't look for it.
The sign.py logic that looks for a cert expiration date in the signing output,
and warns if it doesn't find it, dates back to the old signing tool. Since the
new AzureSignTool doesn't produce that output, every build since the
changeover has been spamming the build output with warnings about not finding
the cert expration date.