Make sign-pkg-{windows,mac}/action.yaml accept lists of files.
With viewer project/watchdog, we need to sign a new executable on each
platform, rubbing our nose (as it were) in the awkwardness of parallel
maintenance to a viewer branch and the viewer-build-util repo shared by all
viewer branches.

Since we must update viewer-build-util for this anyway, make both action.yaml
scripts accept new parameters with lists of files to sign. Give them defaults
matching the previously baked-in lists, so we can promote viewer-build-util
even before all viewer branches explicitly pass those lists.

That lets us override the lists of files to sign on the viewer-watchdog branch
to include the new watchdog executable on each platform.